+++ products/products_test.go
@@ -109,7 +109,7 @@ func TestFutureSettlement(t *testing.T) {
 		// Use debug function to update the settlement price as if from a Oracle
 		f.SetSettlementPrice(ctx, "prices.ETH.value", param.settlementPrice)
 		ep := num.NewUint(param.entryPrice)
-		fa, _, err := prod.Settle(ep, num.DecimalFromInt64(param.position))
+		fa, _, err := prod.Settle(ep, num.NewUint(1), num.DecimalFromInt64(param.position))
 		assert.NoError(t, err)
 		assert.EqualValues(t, param.result, fa.Amount.Uint64())
 	}